move main from controller server
